Start of decoder splitting
