+++ rotkehlchen/chain/evm/decoding/constants.py
@@ -1,0 +1,1 @@
+CPT_GAS = 'gas'
