+++ rotkehlchen/chain/evm/decoding/interfaces.py
@@ -5,16 +5,17 @@ from rotkehlchen.types import ChecksumEvmAddress
 
 if TYPE_CHECKING:
     from rotkehlchen.accounting.structures.base import HistoryBaseEntry
-    from rotkehlchen.chain.ethereum.decoding.base import BaseDecoderTools
-    from rotkehlchen.chain.ethereum.node_inquirer import EthereumInquirer
+    from rotkehlchen.chain.evm.node_inquirer import EvmNodeInquirer
     from rotkehlchen.user_messages import MessagesAggregator
+
+    from .base import BaseDecoderTools
 
 
 class DecoderInterface(metaclass=ABCMeta):
 
     def __init__(
             self,
-            ethereum_inquirer: 'EthereumInquirer',  # pylint: disable=unused-argument
+            evm_inquirer: 'EvmNodeInquirer',  # pylint: disable=unused-argument
             base_tools: 'BaseDecoderTools',  # pylint: disable=unused-argument
             msg_aggregator: 'MessagesAggregator',
     ) -> None:
