+++ rotkehlchen/chain/evm/decoding/structures.py
