+++ rotkehlchen/chain/evm/node_inquirer.py
@@ -198,7 +198,7 @@ class EvmNodeInquirer(LockableQueryMixIn, metaclass=ABCMeta):
         self.web3_mapping: Dict[NodeName, Web3] = {}
         self.rpc_timeout = rpc_timeout
         self.chain_id = blockchain.to_chain_id()
-        self.chain_name = self.chain_id.serialize()
+        self.chain_name = self.blockchain.name.lower()
         self.contract_scan = contract_scan
         self.contract_multicall = contract_multicall
         self.contract_multicall_2 = contract_multicall_2
