+++ rotkehlchen/history/events.py
@@ -18,7 +18,7 @@ from rotkehlchen.exchanges.data_structures import AssetMovement, MarginPosition,
 from rotkehlchen.exchanges.manager import SUPPORTED_EXCHANGES, ExchangeManager
 from rotkehlchen.fval import FVal
 from rotkehlchen.logging import RotkehlchenLogsAdapter
-from rotkehlchen.types import EXTERNAL_LOCATION, Location, Timestamp
+from rotkehlchen.types import EXTERNAL_LOCATION, Location, SupportedBlockchain, Timestamp
 from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.misc import timestamp_to_date
 
@@ -26,7 +26,6 @@ if TYPE_CHECKING:
     from rotkehlchen.accounting.ledger_actions import LedgerAction
     from rotkehlchen.accounting.mixins.event import AccountingEventMixin
     from rotkehlchen.chain.aggregator import ChainsAggregator
-    from rotkehlchen.chain.ethereum.decoding.decoder import EVMTransactionDecoder
     from rotkehlchen.db.dbhandler import DBHandler
     from rotkehlchen.db.drivers.gevent import DBCursor
 
@@ -54,7 +53,6 @@ class EventsHistorian:
             msg_aggregator: MessagesAggregator,
             exchange_manager: ExchangeManager,
             chains_aggregator: 'ChainsAggregator',
-            eth_tx_decoder: 'EVMTransactionDecoder',
     ) -> None:
 
         self.msg_aggregator = msg_aggregator
@@ -62,7 +60,6 @@ class EventsHistorian:
         self.db = db
         self.exchange_manager = exchange_manager
         self.chains_aggregator = chains_aggregator
-        self.eth_tx_decoder = eth_tx_decoder
         self._reset_variables()
 
     def timestamp_to_date(self, timestamp: Timestamp) -> str:
@@ -334,6 +331,7 @@ class EventsHistorian:
             step = self._increase_progress(step, total_steps)
 
         self.processing_state_name = 'Querying ethereum transactions history'
+        ethereum = self.chains_aggregator.get_chain_manager(SupportedBlockchain.ETHEREUM)
         tx_filter_query = ETHTransactionsFilterQuery.make(
             limit=None,
             offset=None,
@@ -343,7 +341,7 @@ class EventsHistorian:
             to_ts=end_ts,
         )
         try:
-            _, _ = self.eth_tx_decoder.transactions.query(
+            _, _ = ethereum.transactions.query(
                 filter_query=tx_filter_query,
                 has_premium=True,  # ignore limits here. Limit applied at processing
                 only_cache=False,
@@ -358,11 +356,11 @@ class EventsHistorian:
         step = self._increase_progress(step, total_steps)
 
         self.processing_state_name = 'Querying ethereum transaction receipts'
-        self.eth_tx_decoder.transactions.get_receipts_for_transactions_missing_them()
+        ethereum.transactions.get_receipts_for_transactions_missing_them()
         step = self._increase_progress(step, total_steps)
 
         self.processing_state_name = 'Decoding raw transactions'
-        self.eth_tx_decoder.get_and_decode_undecoded_transactions(limit=None)
+        ethereum.transactions_decoder.get_and_decode_undecoded_transactions(limit=None)
         step = self._increase_progress(step, total_steps)
 
         # Include all external trades and trades from external exchanges
