+++ rotkehlchen/tests/unit/decoders/test_convex.py
@@ -10,19 +10,19 @@ from rotkehlchen.accounting.structures.base import (
 )
 from rotkehlchen.assets.asset import EvmToken
 from rotkehlchen.chain.ethereum.constants import ZERO_ADDRESS
-from rotkehlchen.chain.ethereum.decoding import EVMTransactionDecoder
-from rotkehlchen.chain.ethereum.decoding.constants import CPT_GAS
+from rotkehlchen.chain.ethereum.decoding.decoder import EthereumTransactionDecoder
 from rotkehlchen.chain.ethereum.modules.convex.constants import CONVEX_POOLS, CPT_CONVEX
 from rotkehlchen.chain.ethereum.modules.convex.decoder import BOOSTER
 from rotkehlchen.chain.ethereum.types import string_to_evm_address
 from rotkehlchen.chain.evm.contracts import EvmContract
+from rotkehlchen.chain.evm.decoding.constants import CPT_GAS
 from rotkehlchen.chain.evm.structures import EvmTxReceipt, EvmTxReceiptLog
 from rotkehlchen.constants import ONE, ZERO
 from rotkehlchen.constants.assets import A_CRV, A_CVX, A_ETH
 from rotkehlchen.constants.ethereum import EthereumConstants
-from rotkehlchen.db.ethtx import DBEthTx
+from rotkehlchen.db.evmtx import DBEvmTx
 from rotkehlchen.fval import FVal
-from rotkehlchen.types import EvmTransaction, Location, deserialize_evm_tx_hash
+from rotkehlchen.types import ChainID, EvmTransaction, Location, deserialize_evm_tx_hash
 from rotkehlchen.utils.hexbytes import hexstring_to_bytes
 from rotkehlchen.utils.misc import hex_or_bytes_to_address
 
@@ -86,13 +86,13 @@ def test_convex_pools(ethereum_inquirer):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0xC960338B529e0353F570f62093Fd362B8FB55f0B')]])  # noqa: E501
-def test_booster_deposit(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_booster_deposit(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x8f643dc245ce64085197692ed98309a94fd176a1e7394e8967ae7bfa10ad1f8f'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0xC960338B529e0353F570f62093Fd362B8FB55f0B')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -106,6 +106,7 @@ def test_booster_deposit(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -133,16 +134,15 @@ def test_booster_deposit(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -179,8 +179,7 @@ def test_booster_deposit(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x53913A03a065f685097f8E8f40284D58016bB0F9')]])  # noqa: E501
-def test_booster_withdraw(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_booster_withdraw(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x79fcbafa4367e0563d3e614f774c5e4257c4e41f124ae8288980a310e2b2b547'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x53913A03a065f685097f8E8f40284D58016bB0F9')
@@ -199,6 +198,7 @@ def test_booster_withdraw(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -236,16 +236,15 @@ def test_booster_withdraw(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -296,13 +295,13 @@ def test_booster_withdraw(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0xFb305A40Dac406BdCF3b85F6311e5430770f44bA')]])  # noqa: E501
-def test_cvxcrv_get_reward(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvxcrv_get_reward(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x5e62ce39159fcdf528905d044e5387c8f21a1eca015d08cebc652bfb9c183611'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0xFb305A40Dac406BdCF3b85F6311e5430770f44bA')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=1655675488,
         block_number=14998088,
         from_address=user_address,
@@ -316,6 +315,7 @@ def test_cvxcrv_get_reward(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -371,16 +371,15 @@ def test_cvxcrv_get_reward(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -445,13 +444,13 @@ def test_cvxcrv_get_reward(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0xe81FC42336c9314A9Be1EDB3F50eA9e275C93df3')]])  # noqa: E501
-def test_cvxcrv_withdraw(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvxcrv_withdraw(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x0a804804cc62f615b72dff55e8c245d9b69aa8f8ed3de549101ae128a4ae432b'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0xe81FC42336c9314A9Be1EDB3F50eA9e275C93df3')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -465,6 +464,7 @@ def test_cvxcrv_withdraw(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -500,16 +500,16 @@ def test_cvxcrv_withdraw(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -546,13 +546,13 @@ def test_cvxcrv_withdraw(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x2AcEcBF2Ee5BFc8eed599D58835EE9A7c45F3E2c')]])  # noqa: E501
-def test_cvxcrv_stake(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvxcrv_stake(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x3cc0b25887e2f0dac7f86fabd81aaafb1e041e84dbe8167885073c443320ad5f'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x2AcEcBF2Ee5BFc8eed599D58835EE9A7c45F3E2c')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -566,6 +566,7 @@ def test_cvxcrv_stake(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -602,16 +603,16 @@ def test_cvxcrv_stake(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -662,13 +663,13 @@ def test_cvxcrv_stake(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x5B186c93A50D3CB435fE2933427d36E6Dc688e4b')]])  # noqa: E501
-def test_cvx_stake(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvx_stake(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0xc33246acb86798b81fe650061061d32751c53879d46ece6991fb4a3eda808103'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x5B186c93A50D3CB435fE2933427d36E6Dc688e4b')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -682,6 +683,7 @@ def test_cvx_stake(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -718,16 +720,16 @@ def test_cvx_stake(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -778,13 +780,13 @@ def test_cvx_stake(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x95c5582D781d507A084c9E5f885C77BabACf8EeA')]])  # noqa: E501
-def test_cvx_get_reward(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvx_get_reward(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0xdaead2f96859462b5800584ecdcf30f2b83a1ca2c36c49a838b23e43c61d803f'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = '0x95c5582D781d507A084c9E5f885C77BabACf8EeA'
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -798,6 +800,7 @@ def test_cvx_get_reward(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -864,16 +867,16 @@ def test_cvx_get_reward(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -910,13 +913,13 @@ def test_cvx_get_reward(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x84BCE169c271e1c1777715bb0dd38Ad9e6381BAa')]])  # noqa: E501
-def test_cvx_withdraw(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_cvx_withdraw(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0xe725bd6e00b840f4fb8f73cd7286bfa18b04a24ca9278cac7249218ee9f420a8'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x84BCE169c271e1c1777715bb0dd38Ad9e6381BAa')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -930,6 +933,7 @@ def test_cvx_withdraw(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -956,16 +960,16 @@ def test_cvx_withdraw(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -1002,13 +1006,13 @@ def test_cvx_withdraw(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x999EcCEa3C4f9219B1B1B42b4830e62c26004B40')]])  # noqa: E501
-def test_claimzap_abracadabras(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_claimzap_abracadabras(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0xe03d27127fda879144ea4cc587470bd37040be9921ff6a90f48d4efd0cb4fe13'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x999EcCEa3C4f9219B1B1B42b4830e62c26004B40')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -1022,6 +1026,7 @@ def test_claimzap_abracadabras(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -1039,16 +1044,16 @@ def test_claimzap_abracadabras(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
+
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
@@ -1085,13 +1090,13 @@ def test_claimzap_abracadabras(database, ethereum_manager, eth_transactions):
 
 
 @pytest.mark.parametrize('ethereum_accounts', [[string_to_evm_address('0x0C3Cc503EaE928Ed6B5b01B8a9EE8de2855d03Ac')]])  # noqa: E501
-def test_claimzap_cvx_locker(database, ethereum_manager, eth_transactions):
-    msg_aggregator = database.msg_aggregator
+def test_claimzap_cvx_locker(database, ethereum_inquirer, eth_transactions):
     tx_hex = '0x53e092e6f25e540d6323af851a1e889276096d58ec25495aef4500467ef2753c'
     evmhash = deserialize_evm_tx_hash(tx_hex)
     user_address = string_to_evm_address('0x0C3Cc503EaE928Ed6B5b01B8a9EE8de2855d03Ac')
     transaction = EvmTransaction(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         timestamp=0,
         block_number=0,
         from_address=user_address,
@@ -1105,6 +1110,7 @@ def test_claimzap_cvx_locker(database, ethereum_manager, eth_transactions):
     )
     receipt = EvmTxReceipt(
         tx_hash=evmhash,
+        chain_id=ChainID.ETHEREUM,
         contract_address=None,
         status=True,
         type=0,
@@ -1132,16 +1138,15 @@ def test_claimzap_cvx_locker(database, ethereum_manager, eth_transactions):
             ),
         ],
     )
-    dbethtx = DBEthTx(database)
-    with dbethtx.db.user_write() as cursor:
-        dbethtx.add_ethereum_transactions(cursor, [transaction], relevant_address=None)
-    decoder = EVMTransactionDecoder(
+    dbevmtx = DBEvmTx(database)
+    with dbevmtx.db.user_write() as cursor:
+        dbevmtx.add_evm_transactions(cursor, [transaction], relevant_address=None)
+    decoder = EthereumTransactionDecoder(
         database=database,
-        ethereum_manager=ethereum_manager,
+        ethereum_inquirer=ethereum_inquirer,
         transactions=eth_transactions,
-        msg_aggregator=msg_aggregator,
     )
-    with dbethtx.db.user_write() as cursor:
+    with dbevmtx.db.user_write() as cursor:
         events = decoder.decode_transaction(cursor, transaction=transaction, tx_receipt=receipt)
     expected_events = [
         HistoryBaseEntry(
